# swagger.yaml
openapi: 3.0.4
info:
  title: Paved Way API
  description: API per la gestione degli eventi sulla mobilità della città di Trento.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Server locale per lo sviluppo e il test.
...
paths:
...
  /users:
    get:
      tags:
        - Utenti
      summary: Restituisce la lista degli utenti
      description: >-
        Restituisce la lista degli utenti registrati. 
        Devi essere autenticato come amministratore per vedere la lista degli utenti.
      security:
        - ApiKeyAuth: []
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
          description: Numero di utenti da saltare prima di restituire i risultati.
        - in: query
          name: limit
          schema:
            type: integer
          description: Numero massimo di utenti da restituire.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserInfo"
        "401":
          description: Non autorizzato. Devi fornire una chiave API valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Accesso negato. Devi essere un amministratore per vedere la lista degli utenti.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    post:
      tags:
        - Utenti
      summary: Crea un nuovo utente
      description: >-
        Permette la registrazione di un nuovo utente.
        Solo un amministratore può creare un nuovo utente amministratore. 
        L'amministratore deve fornire nome utente, nome, cognome, password, ruolo 
        e email o ufficio in base al tipo di utente (rispettivamente cittadino o amministratore).
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/Citizen"
                - $ref: "#/components/schemas/Admin"
      responses:
        "201":
          description: Utente creato con successo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    allOf:
                      - $ref: "#/components/schemas/UserInfo"
                    description: Informazioni sull'utente creato. Include nome utente, nome, cognome, ruolo e email o ufficio.
        "400":
          description: Richiesta non valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Non autorizzato. Devi fornire una chiave API valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Accesso negato. Devi essere un amministratore per creare un utente amministratore.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Nome utente o email già in uso.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}:
    delete:
      tags:
        - Utenti
      summary: Elimina un utente specifico
      description: >-
        Permette di eliminare un utente specifico in base all'ID fornito. 
        Devi essere autenticato come amministratore per eliminare un utente o essere l'utente stesso.
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID dell'utente da eliminare.
      responses:
        "204":
          description: Utente eliminato con successo.
        "401":
          description: Non autorizzato. Devi fornire una chiave API valida.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Accesso negato. Devi essere un amministratore per eliminare un altro utente.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Utente non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/{id}/reports:
    get:
      tags:
        - Utenti
      summary: Restituisce la lista delle segnalazioni di un utente
      description: Restituisce la lista delle segnalazioni di un utente specifico in base all'ID fornito.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID dell'utente di cui restituire le segnalazioni.
        - in: query
          name: latitude
          schema:
            type: float
          description: Latitudine del punto attorno a cui si vuole ricercare dei cantieri.
        - in: query
          name: longitude
          schema:
            type: float
          description: Longitudine del punto attorno a cui si vuole ricercare dei cantieri.
        - in: query
          name: radius
          schema:
            type: integer
          description: Raggio del cerchio con centro il punto attorno a cui si vuole ricercare dei cantieri.
        - in: query
          name: offset
          schema:
            type: integer
          description: Numero di segnalazioni da saltare prima di restituire i risultati.
        - in: query
          name: limit
          schema:
            type: integer
          description: Numero massimo di segnalazioni da restituire.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Report"
        "404":
          description: Utente non trovato.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Errore interno del server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Chiave API per l'autenticazione.
        Deve essere fornita nell'intestazione della richiesta.
        Deve essere un token JWT valido.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Messaggio di errore.
        code:
          type: integer
          description: Codice di errore.
        message:
          type: string
          description: Messaggio di errore dettagliato.
    User:
      type: object
      required:
        - username
        - name
        - surname
        - password
      properties:
        id:
          type: integer
          description: ID dell'utente (generato automaticamente dal server).
        username:
          type: string
          description: Nome utente.
        name:
          type: string
          description: Nome dell'utente.
        surname:
          type: string
          description: Cognome dell'utente.
        password:
          type: string
          description: Password dell'utente.
        role:
          type: string
          enum:
            - citizen
            - admin
          description: Ruolo dell'utente (cittadino o amministratore, assegnato automaticamente alla registrazione).
    Citizen:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required:
            - email
          properties:
            email:
              type: string
              format: email
              description: Indirizzo email del cittadino.
    Admin:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required:
            - office
          properties:
            office:
              type: string
              description: ufficio di competenza dell'amministratore.
    UserInfo:
      type: object
      required:
        - username
        - name
        - surname
        - role
      properties:
        username:
          type: string
          description: Nome utente.
        name:
          type: string
          description: Nome dell'utente.
        surname:
          type: string
          description: Cognome dell'utente.
        role:
          type: string
          enum:
            - citizen
            - admin
          description: Ruolo dell'utente.
      oneOf:
        - required: [email]
          properties:
            email:
              type: string
              format: email
              description: Indirizzo email del cittadino.
        - required: [office]
          properties:
            office:
              type: string
              description: Ufficio di competenza dell'amministratore.
    Location:
      type: object
      required:
        - latitude
        - longitude
        - street
        - number
        - city
        - code
      properties:
        latitude:
          type: number
          format: float
          description: Latitudine della posizione.
          minimum: -90
          maximum: 90
        longitude:
          type: number
          format: float
          description: Longitudine della posizione.
          minimum: -180
          maximum: 180
        street:
          type: string
          description: Nome della strada.
        number:
          type: string
          description: Numero civico.
        city:
          type: string
          description: Nome della città.
        code:
          type: string
          minLength: 5
          maxLength: 5
          description: Codice di Avviamento Postale.
    Duration:
      type: object
      required:
        - start
      properties:
        start:
          type: string
          format: date-time
          description: Data e ora di inizio dell'evento.
        end:
          type: string
          format: date-time
          description: Data e ora di fine dell'evento.
    Comment:
      type: object
      required:
        - text
      properties:
        id:
          type: integer
          description: ID del commento (generato automaticamente dal server).
        userId:
          type: integer
          description: ID dell'utente che ha scritto il commento.
        text:
          type: string
          description: Testo del commento.
        createdAt:
          type: string
          format: date-time
          description: Data e ora di creazione del commento.
    Event:
      type: object
      required:
        - name
        - info
        - location
        - duration
      properties:
        id:
          type: integer
          description: ID dell'evento (generato automaticamente dal server).
        name:
          type: string
          description: Nome dell'evento.
        info:
          type: string
          description: Descrizione dell'evento.
        location:
          allOf:
            - $ref: "#/components/schemas/Location"
          description: Posizione dell'evento.
        duration:
          allOf:
            - $ref: "#/components/schemas/Duration"
          description: Durata dell'evento.
        createdAt:
          type: string
          format: date-time
          description: Data e ora di creazione dell'evento (ora attuale se non specificata).
        comments:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
          description: Lista dei commenti associati all'evento.
    Report:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            createdBy:
              type: string
              description: ID dell'utente che ha creato la segnalazione.
            photos:
              type: array
              items:
                type: string
                format: uri
              description: Lista degli URL delle foto associate alla segnalazione.
            rating:
              type: integer
              description: Punteggio della segnalazione (impostato a 0 per una nuova segnalazione).
            status:
              type: string
              enum:
                - pending
                - approved
                - rejected
                - solved
              description: Stato della segnalazione.
...